stats header now calculates in its respective pages
