+++ src/components/Admin/AdminHSK.tsx
@@ -11,6 +11,10 @@ import RoomHeader from "./RoomHeader.tsx";
 const AdminHSK = () => {
   const [HSKrooms, setHSKrooms] = useState<any[]>([]);
   const [sortedRooms, setSortedRooms] = useState<any[]>([]);
+  const [stats, setStats] = useState({
+    totalHousekeepers: 0,
+    totalRoomsToClean: 0,
+  });
 
   useEffect(() => {
     const roomsCollectionRef = collection(db, "AdminHSK");
@@ -19,6 +23,25 @@ const AdminHSK = () => {
       const dirtyRooms = allRooms.filter(
         (room) => room.roomStatus === "Dirty" || room.roomStatus === "ON CHANGE"
       );
+      const maxHousekeeper = dirtyRooms.reduce((max, room) => {
+        if (room.assignedtoHSK && room.assignedtoHSK.startsWith("HSK")) {
+          const num = parseInt(room.assignedtoHSK.replace("HSK", ""), 10);
+          return num > max ? num : max;
+        } else {
+          console.log("maxHousekeeper can't find a room assigned to HSK");
+        }
+        return max;
+      }, 0);
+      const HSKfilterrooms = allRooms.filter(
+        ((room) => room.roomStatus === "Dirty" || room.roomStatus === "ON CHANGE")
+      ).length;
+
+      const updatedStats = {
+        totalHousekeepers: maxHousekeeper,
+        totalRoomsToClean: HSKfilterrooms,
+      };
+
+      setStats(updatedStats);
       setHSKrooms(dirtyRooms);
       setSortedRooms(dirtyRooms);
     });
@@ -30,7 +53,7 @@ const AdminHSK = () => {
       <AdminNavbar />
       <div className="dashboard-header flex justify-between items-center m-0 mb-4">
         <h1 className="text-3xl text-wine">Housekeepers</h1>
-        <StatsHeader pagename="AdminHSK" displayedRooms={HSKrooms} />
+        <StatsHeader pagename="AdminHSK" stats={stats} />
       </div>
       <SortButton rooms={HSKrooms} onSortedRooms={setSortedRooms} />
       <div className="room-header">
