+++ src/components/Admin/AdminSUP.tsx
@@ -11,6 +11,11 @@ import RoomHeader from "./RoomHeader.tsx";
 const AdminSUP = () => {
   const [SUProoms, setSUProoms] = useState<any[]>([]);
   const [sortedRooms, setSortedRooms] = useState<any[]>([]);
+  const [stats, setStats] = useState({
+    totalSupervisors: 0,
+    totalRoomsToInspect: 0,
+  });
+
 
   useEffect(() => {
     const roomsCollectionRef = collection(db, "AdminHSK");
@@ -19,6 +24,26 @@ const AdminSUP = () => {
       const cleanRooms = allRooms.filter(
         (room) => room.roomStatus === "Clean" || room.coStatus === "INSPECTED"
       );
+      const maxSupervisor = cleanRooms.reduce((max, room) => {
+        if (room.assignedtoSUP && room.assignedtoSUP.startsWith("SUP")) {
+          const num = parseInt(room.assignedtoSUP.replace("SUP", ""), 10);
+          return num > max ? num : max;
+        } else {
+          console.log("maxSupervisor can't find a room assigned to SUP");
+        }
+        return max;
+      }, 0);
+      
+      const SUPfilterrooms = allRooms.filter(
+        (room) => room.roomStatus === "Clean"
+      ).length;
+
+      const updatedStats = {
+        totalSupervisors: maxSupervisor,
+        totalRoomsToInspect: SUPfilterrooms,
+      };
+
+      setStats(updatedStats);
       setSUProoms(cleanRooms);
       setSortedRooms(cleanRooms);
     });
@@ -30,7 +55,7 @@ const AdminSUP = () => {
       <AdminNavbar />
       <div className="dashboard-header flex justify-between items-center m-0 mb-4">
         <h1 className="text-3xl text-wine">Supervisors</h1>
-        <StatsHeader pagename="AdminSUP" displayedRooms={SUProoms} />
+        <StatsHeader pagename="AdminSUP" stats={stats} />
       </div>
       <SortButton rooms={SUProoms} onSortedRooms={setSortedRooms} />
       <div className="room-header">
